move toCustomField and toCustomProperties to the utils module
